Refactor options
- Split the `config.ts` in smaller files, one for each
  interface.
- Expose interface directly from the main module instead of
  the `config` namespace.
- Rename `DEFAULT_PROJECT_OPTIONS` to `DEFAULT_PROJECT`.
