+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ## Next
 
+- **[Breaking]** Expose option interfaces directly in the main module instead of the `config` namespace.
+- **[Breaking]** Rename `DEFAULT_PROJECT_OPTIONS` to `DEFAULT_PROJECT`.
 - **[Internal]** Convert gulpfile to Typescript, use `ts-node` to run it.
 - **[Internal]** Update dependencies
