+++ src/lib/project-tasks/index.ts
@@ -1,5 +1,5 @@
 import {Gulp} from "gulp";
-import {ProjectOptions} from "../config/config";
+import {Project} from "../project";
 import * as bumpMajor from "./bump-major";
 import * as bumpMinor from "./bump-minor";
 import * as bumpPatch from "./bump-patch";
@@ -7,7 +7,7 @@ import * as lint from "./lint";
 import * as lintFix from "./lint-fix";
 import * as tslintJson from "./tslint-json";
 
-export function registerAll(gulp: Gulp, project: ProjectOptions) {
+export function registerAll(gulp: Gulp, project: Project) {
   bumpMajor.registerTask(gulp, project);
   bumpMinor.registerTask(gulp, project);
   bumpPatch.registerTask(gulp, project);
