+++ src/lib/target-generators/webpack.ts
@@ -3,7 +3,8 @@ import {FSWatcher} from "fs";
 import {Gulp} from "gulp";
 import {TaskFunc} from "orchestrator";
 import {posix as path} from "path";
-import {ProjectOptions, WebpackTarget} from "../config/config";
+import {Project} from "../project";
+import {WebpackTarget} from "../targets";
 import * as buildTypescript from "../task-generators/build-typescript";
 import * as buildWebpack from "../task-generators/build-webpack";
 import * as clean from "../task-generators/clean";
@@ -32,7 +33,7 @@ interface Locations {
  * @param target
  * @returns {Locations} The absolute locations
  */
-function resolveLocations(project: ProjectOptions, target: WebpackTarget): Locations {
+function resolveLocations(project: Project, target: WebpackTarget): Locations {
   const targetDir: string = target.targetDir === undefined ? target.name : target.targetDir;
 
   const rootDir: string = toUnix(project.root);
@@ -44,7 +45,7 @@ function resolveLocations(project: ProjectOptions, target: WebpackTarget): Locat
   return {rootDir, srcDir, webpackDir, buildDir, distDir};
 }
 
-export function generateTarget(gulp: Gulp, project: ProjectOptions, target: WebpackTarget) {
+export function generateTarget(gulp: Gulp, project: Project, target: WebpackTarget) {
   const targetName: string = target.name;
   const locations: Locations = resolveLocations(project, target);
