+++ src/test/index.spec.ts
@@ -1,7 +1,6 @@
 import {assert} from "chai";
 
 describe("WebBuildTools", function () {
-  it("should run at least one test...", function () {
-    assert.equal(2, 1 + 1);
+  it("should compile", function () {
   });
 });
